+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - DataTable => Datatable
 - default_order
 - datatable id: consistency btw ajax_search() and render => settings
-  
+- rspec: ajax search
+
 ### new feature
 - embed youtube in score (or just link to search page ?)
 - sort in list table